удаление комментариев. модульность.
