+++ index.html
@@ -177,6 +177,7 @@
     </section>
   </template>
   <!-- Подключение скрипта -->
-  <script src="js/main.js"></script>
+  <script src="js/main.js" type="module"></script>
 </body>
+
 </html>
