+++ js/data.js
@@ -1,0 +1,38 @@
+const ID_COUNT_MAX = 25;
+
+const LikesCount = {
+  MIN: 15, // минимальное значение лайков
+  MAX: 200, //максимальное значение лайков
+};
+
+const CommentsCount = {
+  MIN: 0, // минимальное значение комментариев
+  MAX: 200, //максимальное значение комментариев
+};
+
+const COMMENTS = [
+  'Всё отлично!',
+  'В целом всё неплохо. Но не всё.',
+  'Когда вы делаете фотографию, хорошо бы убирать палец из кадра. В конце концов это просто непрофессионально.',
+  'Моя бабушка случайно чихнула с фотоаппаратом в руках и у неё получилась фотография лучше.',
+  'Я поскользнулся на банановой кожуре и уронил фотоаппарат на кота и у меня получилась фотография лучше.',
+  'Лица у людей на фотке перекошены, как будто их избивают. Как можно было поймать такой неудачный момент?!',
+];
+
+const DESCRIPTIONS = [
+  'Если смогу, я сделаю это. Конец истории.',
+  'Смейтесь как только умеете, любите столько, сколько живете.',
+  'Помните: вы единственный человек, который может наполнить ваш мир солнечным светом.',
+  'Я полностью уверена, что я — диснеевская принцесса, которую еще не придумали.',
+  'Жизнь — это всего лишь серия крошечных чудес, поэтому обратите внимание на них.',
+  'Ни о чем не беспокойтесь. Потому что все лучшие умы на работе.',
+  'Всегда начинайте свой день с хороших людей и кофе.',
+  'Моя жизнь меняется, потому что меняю ее я.',
+  'Не позволяйте кому-то затушить ваши искры только потому, что их свет сияет в чьих-то глазах.',
+  'Делайте в вашей жизни то, что меньше заставляет вас смотреть в свой телефон.',
+  'Улыбка — единственный тренд в моде, который актуален всегда. ',
+  'Никогда не ищите свое счастье там, где вы его однажды потеряли.',
+  'Жизнь похожа на фотокамеру: вам просто нужно смотреть на нее с улыбкой.',
+];
+
+export { DESCRIPTIONS, COMMENTS, CommentsCount, LikesCount, ID_COUNT_MAX };
